requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
Pillow>=10.0.0
SQLAlchemy==2.0.21
PyMySQL==1.1.0
bcrypt==4.1.2
PyJWT==2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
ultralytics>=8.2.0
